+++ lib/cached/md5_cached.cmake
@@ -1,4 +1,4 @@
 set(lexer_lxx_md5_cached "463d007a1f237dd01cb3020208efda3c")
-set(parser_yxx_md5_cached "a1a3ad0d5717f0cc50417099ff2ae02f")
+set(parser_yxx_md5_cached "a49d72232e30c06667be4eb862bc67eb")
 set(regex_lexer_lxx_md5_cached "8906a52bfa0c5ae26354cb272348e656")
 set(regex_parser_yxx_md5_cached "68ec070becef5e161c3b97d085b0810e")